Move OpenID-related handler stuff to own package
