+++ handlers/openid/openid.go
@@ -1,0 +1,39 @@
+package openid
+
+import (
+	"encoding/json"
+	"github.com/vouch/vouch-proxy/handlers/common"
+	"github.com/vouch/vouch-proxy/pkg/cfg"
+	"github.com/vouch/vouch-proxy/pkg/structs"
+	"golang.org/x/oauth2"
+	"io/ioutil"
+	"net/http"
+)
+
+var (
+	log = cfg.Cfg.Logger
+)
+
+func GetUserInfoFromOpenID(client *http.Client, user *structs.User, customClaims *structs.CustomClaims, ptoken *oauth2.Token) (rerr error) {
+	userinfo, err := client.Get(cfg.GenOAuth.UserInfoURL)
+	if err != nil {
+		return err
+	}
+	defer func() {
+		if err := userinfo.Body.Close(); err != nil {
+			rerr = err
+		}
+	}()
+	data, _ := ioutil.ReadAll(userinfo.Body)
+	log.Infof("OpenID userinfo body: %s", string(data))
+	if err = common.MapClaims(data, customClaims); err != nil {
+		log.Error(err)
+		return err
+	}
+	if err = json.Unmarshal(data, user); err != nil {
+		log.Error(err)
+		return err
+	}
+	user.PrepareUserData()
+	return nil
+}
